Major Update Fixed SpringBootSec + JWT Autho
FIXED ALOT

+++ src/main/java/tn/esprit/spring/AhmedGuedri/Controllers/UserController.java
@@ -6,6 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import tn.esprit.spring.AhmedGuedri.Services.IPWDService;
@@ -13,6 +14,9 @@ import tn.esprit.spring.AhmedGuedri.Services.IUserService;
 import tn.esprit.spring.AhmedGuedri.entities.HashedPWD;
 import tn.esprit.spring.AhmedGuedri.entities.User;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
+import java.security.Principal;
 import java.util.Date;
 import java.util.List;
 
@@ -40,27 +44,51 @@ public class UserController {
     }
 
     @PutMapping("/updateUser")
-    public String updateUser(@RequestBody User u) {
+    @PreAuthorize("hasRole('USER') or hasRole('PROVIDER') or hasRole('ADMIN')")
+    public String updateUser(HttpServletRequest request,@RequestBody User u) {
+        String email= iUserService.UserVerificationReturnEmail(request);
+        if (email.equals("Token Doesn't Match Authenfied User"))
+            return "Token Doesn't Match Authenfied User";
+
+        if (u.getEmail().equals(email))
         return iUserService.updateUser(u);
+        else
+            return "Email Doesn't Match Authenfied User";
     }
 
 
 
     @DeleteMapping("/deleteUser")
-    public ResponseEntity<User> deleteUser(@RequestParam String email) {
-        iUserService.deleteUser(email);
-        return new ResponseEntity<User>(HttpStatus.OK);
+    @PreAuthorize("hasRole('ADMIN')")
+    public String deleteUser(HttpServletRequest request) {
+        String email= iUserService.UserVerificationReturnEmail(request);
+        if (email.equals("Token Doesn't Match Authenfied User"))
+            return "Token Doesn't Match Authenfied User";
+        return iUserService.deleteUser(email);
+    }
+    @PutMapping("/activateUser")
+    @PreAuthorize("hasRole('ADMIN')")
+    public String ActivateUser(HttpServletRequest request, @RequestParam String emailAdress) {
+        String email= iUserService.UserVerificationReturnEmail(request);
+        if (email.equals("Token Doesn't Match Authenfied User"))
+            return "Token Doesn't Match Authenfied User";
+        return iUserService.deleteUser(emailAdress);
+
 
     }
     //implementing the method RetievePasswordInfo
     @GetMapping("/RetievePasswordInfo")
-    public String RetievePasswordInfo(@RequestParam String email) {
+    public String RetievePasswordInfo(HttpServletRequest request) {
+        String email= iUserService.UserVerificationReturnEmail(request);
+        if (email.equals("Token Doesn't Match Authenfied User"))
+            return "Token Doesn't Match Authenfied User";
         return ipwdService.RetievePasswordInfo(email);
 
     }
     //Implementing VerifyUserToken
     @GetMapping("/VerifyUserToken")
     public String VerifyUserToken(@RequestParam String email, @RequestParam Long token) {
+        System.out.println("Before : "+email+" - "+token);
         return iUserService.VerifyUserToken(email, token);
 
     }
@@ -76,5 +104,6 @@ public class UserController {
         iUserService.Authenticate(email);
 
     }
+
 
 }
